requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cufflinks==0.17.3
Flask==2.0.0
gunicorn==20.1.0
line-bot-sdk==1.8.0
matplotlib==3.4.2
pandas==1.2.4
plotly==4.14.3
plotly-utils==0.0.3
psutil==5.8.0
requests==2.22.0
openpyxl






